Fix SQLite to PostgreSQL migration: boolean types and offline mode
- Change INTEGER to BOOLEAN for is_initial, is_terminal, starts_workflow, enabled columns
- Update seed migrations to use TRUE/FALSE instead of 1/0
- Update Rust code to use bool instead of i32 for boolean parameters
- Add SQLX_OFFLINE=true to npm scripts for offline query validation
- Remove remote crate dependency from server
- Regenerate SQLx offline cache for PostgreSQL

Files changed:
- migrations: 20251231000001, 20251231000003, 20251231000005, 20260107000001, 20260109000001, 20260112000005
- models: automation_rule.rs, kanban_column.rs
- package.json: Added SQLX_OFFLINE=true to backend scripts

+++ crates/local-deployment/src/container.rs
@@ -471,7 +471,7 @@ impl LocalContainerService {
 
             let (exit_code, status) = match status_result {
                 Ok(exit_status) => {
-                    let code = exit_status.code().unwrap_or(-1) as i64;
+                    let code = exit_status.code().unwrap_or(-1);
                     let status = if exit_status.success() {
                         ExecutionProcessStatus::Completed
                     } else {
